Tests/LibWeb/Text/input/wpt-import/encoding-detection/zh-GBK.tentative.html

<!doctype html>
<title>zh GBK</title>
<script src=../resources/testharness.js></script>
<script src=../resources/testharnessreport.js></script>
<p>侶角寧몸俚륜긍쯤꿎桿。</p>
<script>
setup({explicit_done:true});
onload = function() {
    test(function() {
        assert_equals(document.characterSet, "GBK", 'Expected GBK');
    }, "Check detection result");
      done();
};
</script>
